task ownership / authorization needed for tasks handeling

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { Task } from './task.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskStatus } from './task.status.enum';
 import { GetTasksFilterDTO } from './dto/get-tasks-filter.dto';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,16 +15,16 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) {}
 
-    public async getTasks(filterDto: GetTasksFilterDTO): Promise<Task[]>  {
-        return this.taskRepository.getTasks(filterDto);
+    public async getTasks(filterDto: GetTasksFilterDTO, user: User): Promise<Task[]>  {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    public async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    public async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    public async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    public async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
         if (!found) {
             throw new NotFoundException(`task with ID: '${id}' not found`);
@@ -32,17 +33,17 @@ export class TasksService {
         return found;
     }
 
-    public async deleteTaskById(id: number): Promise<void> {
+    public async deleteTaskById(id: number, user: User): Promise<void> {
 
-        const result = await this.taskRepository.delete(id);
+        const result = await this.taskRepository.delete({ id, userId: user.id });
 
         if (result.affected === 0) {
             throw new NotFoundException(`task with ID: '${id}' not found`);
         }
     }
 
-    public async updateTaskById(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    public async updateTaskById(id: number, status: TaskStatus, user): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
